+++ src/Localization/LOCALIZATION_KEY.mjs
@@ -1,5 +1,3 @@
-export const LOCALIZATION_KEY_LANGUAGE = "language";
-
 export const LOCALIZATION_KEY_SYSTEM = "system";
 
 export const LOCALIZATION_KEY_SYSTEM_WITH_LANGUAGE = "system-with-language";