+++ src/StyleSheetManager/StyleSheetManager.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {{addRootStyleSheet: (style_sheet: CSSStyleSheet, beginning: boolean) => Promise<void>, addStyleSheetsToShadow: (shadow: ShadowRoot) => Promise<void>, generateVariablesRootStyleSheet: (prefix: string, variables: {[key: string]: string}, beginning: boolean) => Promise<void>}} StyleSheetManager
- */